Fixes Eigen scalar sum warning
Eigen doesn't like the fact that we're creating an Array
of Bools, and then attempting to sum those booleans up.

Instead, we need to cast the Array into an int, and then
sum over it to store into the sum integer.

Resolves https://github.com/RethinkRobotics-opensource/sns_ik/issues/56

+++ sns_ik_lib/src/sns_ik_math_utils.cpp
@@ -223,7 +223,7 @@ bool pinv_forBarP(const MatrixD &W, const MatrixD &P, MatrixD *inv) {
   MatrixD tmp;
   bool invertible;
 
-  int sizeBarW = (W.diagonal().array() > 0.99).sum();
+  int sizeBarW = (W.diagonal().array() > 0.99).cast<int>().sum();
   MatrixD barW(sizeBarW, W.cols());
   int rowsBarW = 0;
